backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pydantic
pandas
yfinance
pytest
httpx
